NEWS	R Documentation

News for package ‘phylopomp’

Changes in phylopomp version 0.19.2

There are new BDEI and BDSS models.

Changes in phylopomp version 0.19.1

Fixed bug in seirs_pomp.

Change tf argument of parse_newick to time in keeping with usage elsewhere.

Changes in phylopomp version 0.19

Breaking changes in the parameterizations of the SEIR, SIIR, SIR, and Strains models. See the manual pages for details.

The new function geneal_scale allows rescaling and/or shifting of a genealogy's time-scale.

The behavior of the geneal_sum function and the genealogy-addition operator (+) have changed. The sum is now inclusive rather than exclusive: it expands the time-interval as needed. NB: this can have unintended consequences.

There is a new model, MERS.

The SI2R model has been completely changed. It now features two demes, one of which transmits at a rate that is a multiple of the other.

In the SI2R model, the sampling is assumed to be destructive.

The YAML-translator has been improved and updated.

Improvements to unit testing.

Changes in phylopomp version 0.18

The integration of the filter for the SEIRS model has changed.

sir_pomp, seirs_pomp, and strains_pomp now expect initial conditions expressed as fractions of a specified host population size (pop).

When times or branch-lengths are converted to decimal representation (e.g., by newick() or yaml()), more care is taken to preserve precision.

The curtail function now takes arguments t0 and time, consistent with the usage in the rest of the package.

Changes in phylopomp version 0.17

This release contains a relatively large refactorization of the C++ library.

There is a new facility for summing genealogies together. The sum of two genealogies is defined as the union of all the lineages, curtailed to the intersections of their time-intervals.

The default behavior of the newick() command is to generate code in an extended Newick format. Specifically, information on the type and deme of nodes is returned using a phylopomp-specific metadata tag embedded in each label. Optionally, one can generate ordinary Newick-format, in which case all sample demes are represented as tips, the terminal twigs of inline tips having zero length. See the manual pages for details.

The Newick parser, parse_newick() can now read either of these formats. It interprets all encountered tips to be samples, unless otherwise indicated using the metadata tag mentioned above. See the manual pages for details.

lbdp_exact() now gives correct results even for empty genealogies.

Internally, phylopomp now represents the demes more carefully. The deme-number 0 is reserved for the ‘undeme’, i.e., nodes or branches with unspecified deme.

The use of palettes in the plotting routines is now more consistent.

Changes in phylopomp version 0.16

gendat() now returns the deme for samples.

The Strains model (see ?Strains) now has a filtering code: strains_pomp().

The supported linear birth-death-sampling processes now allow for both destructive and non-destructive sampling (see ?LBDP).

Changes in phylopomp version 0.15

New model in which three distinct strains compete for susceptibles from a single pool. See Strains.

New model which is identical to TwoSpecies except in that lineages that terminate are replaced by ghosts.

Bug fix in curtail().

Refactoring of lbdp_exact().

Changes in phylopomp version 0.14

The curtail() function now allows one to curtail a genealogy from the left as well as the right.

The filtering codes for the two-species model are now working.

A new sample/death event-type has been introduced into the genealogy_t type.

Corrections of filtering equations for the SEIR and Two-species models.

Yet another refactoring. The SEIRS, LBDP, and SIRS models have been refactored to streamline the filtering calculations in a generalizable fashion.

Changes in phylopomp version 0.13

Much faster computations with complex genealogies.

New S2I2R2 model of two-species system.

Changes in phylopomp version 0.11

The SEIRS model pomp object has been refactored.

Changes in phylopomp version 0.10

getInfo(), parse_newick() now always trace lineages.

Lineage and deme information has been refactored: nodes no longer carry deme information, balls do.

The genealogy description (getInfo(description=TRUE)) now reports both deme and lineage.

curtail() more carefully rewinds the history.

Parameter name changes in some of the examples.

In plotting, trees are shifted down on the y axis.

Changes in phylopomp version 0.9

Inclusion of pomp methods for SEIRS model.

Refactored sirs_pomp and lbdp_pomp.

The new function geneal extracts the bare genealogy from a simulated master process object.

Changes in the interfaces to lbdp_exact, lbdp_pomp, moran_exact, and sir_pomp. These functions now take as data objects of class ‘gpgen’.

The old function newick2df is no more.

Considerable reworking of the internals.

Definition of new class, ‘gpsim’, for bare genealogies.

Definition of new class, ‘gpyaml’, for YAML descriptions.

Changes in phylopomp version 0.8

The interface to getInfo has changed.

Changes in phylopomp version 0.7

In genealogy diagrams, the nodes are now colored according to deme. This can be controlled through the use of the diagram function's palette argument (cf. plot).

The new function curtail curtails a genealogy (i.e., casts away nodes beyond a given time).

The new function newick extracts a Newick-format description of a genealogy.

The SIRS model is included as a synonym for SIR. Both allow waning of immunity.

The lineages function now returns the saturation and a code for the event-type as well as the lineage-count. The returned data frame is now in a long format.

The newick2df function has been rewritten.

A new function parse_newick allows one to parse a Newick string into the formats used in phylopomp.

Changes in phylopomp version 0.6

The option to generate “non-compact” Newick-format tree representations has been removed: all Newick representations are now in the “compact” format. In this format, each node has a color indicated by a letter immediately following its closing parenthesis: this is ‘b’ if the node holds a blue ball (i.e., it represents a sample) and ‘g’ otherwise. Information about the node's deme and its name are provided as well.

Previously, the Newick formatter employed “invisible” nodes to assist in the formatting of plots. As of this release, they no longer contain such nodes: the generated Newick strings now accurately represent the forest of trees corresponding to the genealogy process. The proper formatting for plots is accomplished within the plotting routines themselves.

Changes in phylopomp version 0.5

Function moran_exact computes the exact likelihood under the Moran model.

Changes in phylopomp version 0.4

Refactored the SIRS model ‘pomp’ object.

The obscure operation now drops single-child internal nodes in addition to erasing all deme information.

Corrected certain inconsistencies in the way the lineage-count table was reported and used in pomp objects. The lbdp_pomp and sir_pomp functions assume a particular form of the lineage count table, such that every entry corresponds to either a root, a branch point, or a sample (but NB, the last entry is treated differently). This assumption was violated beginning with version 0.2.0.0 in that the lineage count table returned by lineages() contained an entry for internal nodes. Now, when obscure=TRUE (the default), the behavior is as it was before, consistent with the assumptions of lbdp_pomp and sir_pomp

Changes in phylopomp version 0.2

Streamlined of the GP definition.

Changes in phylopomp version 0.1

Updated for use with pomp version 4.6.