NEWS	R Documentation

News for package ‘hte’

Changes in pomp version 0.0.3

The fake data in fake_data no longer have contact isolation when patients are out of the hospital.

Changes in pomp version 0.0.2

The codes have been refactored to reduce unnecessary data manipulation.