NEWS	R Documentation

News for package ‘phylopomp’

Changes in phylopomp version 0.5.0

Function moran_exact computes the exact likelihood under the Moran model.

Changes in phylopomp version 0.4.0

Refactored the SIRS model ‘pomp’ object.

The obscure operation now drops single-child internal nodes in addition to erasing all deme information.

Corrected certain inconsistencies in the way the lineage-count table was reported and used in pomp objects. The lbdp_pomp and sir_pomp functions assume a particular form of the lineage count table, such that every entry corresponds to either a root, a branch point, or a sample (but NB, the last entry is treated differently). This assumption was violated beginning with version 0.2.0.0 in that the lineage count table returned by lineages() contained an entry for internal nodes. Now, when obscure=TRUE (the default), the behavior is as it was before, consistent with the assumptions of lbdp_pomp and sir_pomp

Changes in phylopomp version 0.2.0

Streamlined of the GP definition.

Changes in phylopomp version 0.1.0

Updated for use with pomp version 4.6.