NEWS	R Documentation

News for Package 'aakmisc'

Changes in aakmisc version 0.29

Add numbers2words, cadged from Andy Teuchen.

Changes in aakmisc version 0.28

Problem with scinot has been fixed.

aakmisc now requires R version 4 or later.

Changes in aakmisc version 0.27

New matrix2latex function.

Changes in aakmisc version 0.26

The package no longer automatically loads ggplot2, scales, grid, plyr, reshape2, magrittr, and readr.

scinot is now suitable for use in ggplot::scale_X_Y.

The new function trnc allows truncation to the plot window when used in ggplot::scale_X_Y.

Changes in aakmisc version 0.25

The license under which the software is provided has been upgraded to GPL version 3.

Package readr is a new dependency and scales is explicitly imported.

Changes in aakmisc version 0.24

The license under which the software is provided has been upgraded to GPL version 3.

Changes in aakmisc version 0.23

startTunnel now takes a user option.

Changes in aakmisc version 0.21

Codes for lazy-loading knitr caches have been added. These useful functions are shamelessly cadged from Peter DeWitt's qwrap2 package.

Changes in aakmisc version 0.16

bake and stew have been moved to pomp.

Changes in aakmisc version 0.15

New stew function (similar to bake).

Changes in aakmisc version 0.12

On attachment, the default ggplot2 theme is theme_bw().

Changes in aakmisc version 0.11

New bake function.

Changes in aakmisc version 0.10

A new scinot function for formatting numbers in scientific notation.

Changes in aakmisc version 0.9

There is a new plotMatrix method for generating scatterplot matrices and density plots from data-frames.

Changes in aakmisc version 0.8

will attempt to clean up ssh tunnel processes on exit from R