NEWS	R Documentation

News for package ‘hte’

Changes in hte version 0.0.9

Fixed bug in the parameterization of the unit-specific transmission model (trans_unit_spec_objfun). The a parameter is now global.

Changes in hte version 0.0.8

Fixed bug in the parameterization of the homogeneous transmission model (trans_homog_objfun).

Changes in hte version 0.0.7

It is possible to set maxit and tol at the top level.

Changes in hte version 0.0.6

Efficiency is much greater now that patients out of hospital are jumped immediately to the next admission.

Changes in hte version 0.0.5

Transmission model is reparameterized in terms of a and b.

Changes in hte version 0.0.4

Refactor the independent model objective functions.

Changes in hte version 0.0.3

The fake data in fake_data no longer have contact isolation when patients are out of the hospital.

Changes in hte version 0.0.2

The codes have been refactored to reduce unnecessary data manipulation.