NEWS	R Documentation

News for package ‘hte’

Changes in hte version 0.0.21

Unit-specific agent-based simulation. See run_abm.

Changes in hte version 0.0.20

Unit-specific models now have unit-specific values of p0.

Changes in hte version 0.0.19

run_abm has been extended to allow for specified testing of individual patients.

Changes in hte version 0.0.18

A bug in run_abm has been fixed.

Changes in hte version 0.0.17

An agent-based simulator for the transmission model is included (run_abm).

Simulated movement data from codes of RJ Woods are included (sim_pat_mov).

Changes in hte version 0.0.16

Fixed error in transmission-model fixed-point calculation.

Changes in hte version 0.0.15

Testing can convey information about status (parameter eta).

Changes in hte version 0.0.14

The transmission models now contain an intercept term iota.

Changes in hte version 0.0.13

lambda.out is now present as a parameter in all models.

Changes in hte version 0.0.12

The recovery rate gamma is assumed constant everywhere (in and out of hospital) for all models.

Changes in hte version 0.0.11

The filter functions now return the patient ID.

The simulated data in fake_data now use a character variable as the unique identifier for the patients.

Changes in hte version 0.0.10

Export new functions indep_homog_filter, indep_unit_spec_filter, trans_homog_filter, and trans_unit_spec_filter, which perform appropriate versions of the filtering computations.

Changes in hte version 0.0.9

Fixed bug in the parameterization of the unit-specific transmission model (trans_unit_spec_objfun). The a parameter is now global.

Changes in hte version 0.0.8

Fixed bug in the parameterization of the homogeneous transmission model (trans_homog_objfun).

Changes in hte version 0.0.7

It is possible to set maxit and tol at the top level.

Changes in hte version 0.0.6

Efficiency is much greater now that patients out of hospital are jumped immediately to the next admission.

Changes in hte version 0.0.5

Transmission model is reparameterized in terms of a and b.

Changes in hte version 0.0.4

Refactor the independent model objective functions.

Changes in hte version 0.0.3

The fake data in fake_data no longer have contact isolation when patients are out of the hospital.

Changes in hte version 0.0.2

The codes have been refactored to reduce unnecessary data manipulation.