NEWS	R Documentation

News for package ‘phylopomp’

Changes in phylopomp version 0.10

getInfo(), parse_newick() now always trace lineages.

Lineage and deme information has been refactored: nodes no longer carry deme information, balls do.

The genealogy description (getInfo(description=TRUE)) now reports both deme and lineage.

curtail() more carefully rewinds the history.

Parameter name changes in some of the examples.

Changes in phylopomp version 0.9

Inclusion of pomp methods for SEIRS model.

Refactored sirs_pomp and lbdp_pomp.

The new function geneal extracts the bare genealogy from a simulated master process object.

Changes in the interfaces to lbdp_exact, lbdp_pomp, moran_exact, and sir_pomp. These functions now take as data objects of class ‘gpgen’.

The old function newick2df is no more.

Considerable reworking of the internals.

Definition of new class, ‘gpsim’, for bare genealogies.

Definition of new class, ‘gpyaml’, for YAML descriptions.

Changes in phylopomp version 0.8

The interface to getInfo has changed.

Changes in phylopomp version 0.7

In genealogy diagrams, the nodes are now colored according to deme. This can be controlled through the use of the diagram function's palette argument (cf. plot).

The new function curtail curtails a genealogy (i.e., casts away nodes beyond a given time).

The new function newick extracts a Newick-format description of a genealogy.

The SIRS model is included as a synonym for SIR. Both allow waning of immunity.

The lineages function now returns the saturation and a code for the event-type as well as the lineage-count. The returned data frame is now in a long format.

The newick2df function has been rewritten.

A new function parse_newick allows one to parse a Newick string into the formats used in phylopomp.

Changes in phylopomp version 0.6

The option to generate “non-compact” Newick-format tree representations has been removed: all Newick representations are now in the “compact” format. In this format, each node has a color indicated by a letter immediately following its closing parenthesis: this is ‘b’ if the node holds a blue ball (i.e., it represents a sample) and ‘g’ otherwise. Information about the node's deme and its name are provided as well.

Previously, the Newick formatter employed “invisible” nodes to assist in the formatting of plots. As of this release, they no longer contain such nodes: the generated Newick strings now accurately represent the forest of trees corresponding to the genealogy process. The proper formatting for plots is accomplished within the plotting routines themselves.

Changes in phylopomp version 0.5

Function moran_exact computes the exact likelihood under the Moran model.

Changes in phylopomp version 0.4

Refactored the SIRS model ‘pomp’ object.

The obscure operation now drops single-child internal nodes in addition to erasing all deme information.

Corrected certain inconsistencies in the way the lineage-count table was reported and used in pomp objects. The lbdp_pomp and sir_pomp functions assume a particular form of the lineage count table, such that every entry corresponds to either a root, a branch point, or a sample (but NB, the last entry is treated differently). This assumption was violated beginning with version 0.2.0.0 in that the lineage count table returned by lineages() contained an entry for internal nodes. Now, when obscure=TRUE (the default), the behavior is as it was before, consistent with the assumptions of lbdp_pomp and sir_pomp

Changes in phylopomp version 0.2

Streamlined of the GP definition.

Changes in phylopomp version 0.1

Updated for use with pomp version 4.6.