NEWS	R Documentation

News for package ‘circumstance’

Changes in circumstance version 0.0.9

We use doFuture::%dofuture% in preference to foreach::%dopar%.

Changes in circumstance version 0.0.8

New parallel pmcmc methods for parallel particle Markov chain Monte Carlo.

Changes in circumstance version 0.0.7

Information on parallel backend is now stored as an attribute of returned objects.

Changes in circumstance version 0.0.5

New mif2 method for parallel iterated filtering.

New continue method for continuing mif2 computations.

Changes in circumstance version 0.0.4

Use with doFuture is envisioned.

Streamline some of the definitions.

Improve coverage.

Changes in circumstance version 0.0.3

Updates for new version of pomp.

Changes in circumstance version 0.0.1

Improved the parallel pfilter facility.

Changes in circumstance version 0.0.0

A parallel pfilter is provided.